Rename GUI tests
