+++ src/tests/gui/loot_settings_test.h
